DEV: Convert theme-upload modal to component-based API
