+++ app/assets/javascripts/admin/addon/components/theme-upload-add.hbs
@@ -1,0 +1,47 @@
+<DModal
+  class="add-upload-modal"
+  @title={{i18n "admin.customize.theme.add_upload"}}
+  @closeModal={{@closeModal}}
+  @flash={{this.flash}}
+>
+  <:body>
+    <div class="inputs">
+      <section class="field">
+        <input
+          onchange={{this.updateName}}
+          type="file"
+          id="file-input"
+          accept="*"
+        />
+        <label for="file-input">{{i18n
+            "admin.customize.theme.upload_file_tip"
+          }}</label>
+      </section>
+      <section class="field">
+        <label for="theme-variable-name">{{i18n
+            "admin.customize.theme.variable_name"
+          }}</label>
+        <Input id="theme-variable-name" @value={{this.name}} />
+      </section>
+      {{#if this.fileSelected}}
+        {{#if this.errorMessage}}
+          <span class="alert alert-error">{{this.errorMessage}}</span>
+        {{/if}}
+      {{/if}}
+    </div>
+  </:body>
+  <:footer>
+    <DButton
+      @action={{this.upload}}
+      @disabled={{this.disabled}}
+      class="btn btn-primary"
+      @icon="upload"
+      @label="admin.customize.theme.upload"
+    />
+    <DButton
+      class="btn-flat d-modal-cancel"
+      @action={{@closeModal}}
+      @translatedLabel={{i18n "cancel"}}
+    />
+  </:footer>
+</DModal>
