--- app/assets/javascripts/admin/addon/templates/modal/admin-add-upload.hbs
@@ -1,37 +1,0 @@
-<DModalBody @class="add-upload-modal" @title="admin.customize.theme.add_upload">
-  <div class="inputs">
-    <section class="field">
-      <input
-        onchange={{action "updateName"}}
-        type="file"
-        id="file-input"
-        accept="*"
-      />
-      <label for="file-input">{{i18n
-          "admin.customize.theme.upload_file_tip"
-        }}</label>
-    </section>
-    <section class="field">
-      <label for="theme-variable-name">{{i18n
-          "admin.customize.theme.variable_name"
-        }}</label>
-      <Input id="theme-variable-name" @value={{this.name}} />
-    </section>
-    {{#if this.fileSelected}}
-      {{#if this.errorMessage}}
-        <span class="alert alert-error">{{this.errorMessage}}</span>
-      {{/if}}
-    {{/if}}
-  </div>
-</DModalBody>
-
-<div class="modal-footer">
-  <DButton
-    @action={{action "upload"}}
-    @disabled={{this.disabled}}
-    @class="btn btn-primary"
-    @icon="upload"
-    @label="admin.customize.theme.upload"
-  />
-  <DModalCancel @close={{route-action "closeModal"}} />
-</div>
